backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
python-dotenv>=1.0.0
pydantic>=2.10.0
httpx>=0.27.0
colorama>=0.4.6
watchdog>=6.0.0
psutil>=7.0.0

